requirements.txt
alembic==1.7.7
anyio==3.5.0
async-timeout==4.0.2
cachelib==0.6.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
dadata==21.10.1
Deprecated==1.2.13
Flask==2.1.1
Flask-Migrate==3.1.0
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
httpcore==0.14.7
httpx==0.22.0
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.7.0
itsdangerous==2.1.2
Jinja2==3.1.1
Mako==1.2.0
MarkupSafe==2.1.1
packaging==21.3
psycopg2-binary==2.9.3
PyJWT==2.3.0
pyparsing==3.0.7
redis==4.2.1
rfc3986==1.5.0
sniffio==1.2.0
SQLAlchemy==1.4.35
Werkzeug==2.1.1
wrapt==1.14.0
zipp==3.8.0

pytesseract~=0.3.10
Pillow~=9.2.0